requirements.txt
pyfrc==2022.0.0
pynetworktables==2021.0.0
pyntcore==2022.2.1.2
robotpy==2022.0.14
robotpy-hal==2022.2.1.1
robotpy-halsim-gui==2022.2.1.0
robotpy-installer==2022.0.0
robotpy-navx==2022.0.1
robotpy-rev==2022.0.0
robotpy-wpilib-utilities==2022.0.0
robotpy-wpimath==2022.2.1.4
robotpy-wpiutil==2022.2.1.1
wpilib==2022.2.1.5